When deleting a comment, add to display 'deleted' message.

+++ modules/board/board.controller.php
@@ -510,10 +510,19 @@ class boardController extends board
 		// generate comment  controller object
 		$oCommentController = getController('comment');
 
-		$output = $oCommentController->deleteComment($comment_srl, $this->grant->manager);
-		if(!$output->toBool())
+		if($this->module_info->comment_delete_message === 'Y')
 		{
-			return $output;
+			$comment->content = '';
+			$comment->status = 7;
+			$output = $oCommentController->updateCommentByDelete($comment, $this->grant->manager);
+		}
+		else
+		{
+			$output = $oCommentController->deleteComment($comment_srl, $this->grant->manager);
+			if(!$output->toBool())
+			{
+				return $output;
+			}
 		}
 
 		$this->add('mid', Context::get('mid'));
